ml_service/requirements.txt
fastapi
uvicorn[standard]
torch
transformers
datasets
accelerate
sentencepiece
Pillow
pdf2image
pytesseract
opencv-python
